Rename form control files to text input
