+++ components/text_input/text_input.config.js
