+++ fractal_react/__components.js
@@ -5,7 +5,7 @@ module.exports = {
   "css-flexbox-utilities": require("../components/css_flexbox_utilities/css_flexbox_utilities.jsx"),
   "css-spacing-utilities": require("../components/css_spacing_utilities/css_spacing_utilities.jsx"),
   "css-text-utilities": require("../components/css_text_utilities/css_text_utilities.jsx"),
-  "form-control": require("../components/form_control/form_control.jsx"),
+  "form-control": require("../components/text_input/text_input.jsx"),
   "form-group": require("../components/form_group/form_group.jsx"),
   "preview": require("../components/_preview.jsx"),
 };
