+++ styles/_components.scss
@@ -1,3 +1,3 @@
 // Note: order matters!
-@import '../components/form_control/form_control';
+@import '../components/text_input/text_input';
 @import '../components/form_group/form_group';
