+++ styles/phenotypes.css
@@ -167,21 +167,18 @@ a {
   text-decoration: none;
   background-color: transparent;
   -webkit-text-decoration-skip: objects; }
-
-a:hover {
-  color: #008ab3;
-  text-decoration: underline; }
+  a:hover {
+    color: #008ab3;
+    text-decoration: underline; }
 
 a:not([href]):not([tabindex]) {
   color: inherit;
   text-decoration: none; }
-
-a:not([href]):not([tabindex]):focus, a:not([href]):not([tabindex]):hover {
-  color: inherit;
-  text-decoration: none; }
-
-a:not([href]):not([tabindex]):focus {
-  outline: 0; }
+  a:not([href]):not([tabindex]):focus, a:not([href]):not([tabindex]):hover {
+    color: inherit;
+    text-decoration: none; }
+  a:not([href]):not([tabindex]):focus {
+    outline: 0; }
 
 pre,
 code,
@@ -381,9 +378,8 @@ hr {
   border-radius: 3px;
   font-size: 16px;
   width: 100%; }
-
-.FormControl::placeholder {
-  color: rgba(0, 0, 0, 0.41); }
+  .FormControl::placeholder {
+    color: rgba(0, 0, 0, 0.41); }
 
 .FormControl--small {
   font-size: 14px; }
